movendo as funções criar_tableau e solve para classes isoladas
